% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{add_NA_colorlegend}
\alias{add_NA_colorlegend}
\title{Add a \verb{no data} (mask) color to legend}
\usage{
add_NA_colorlegend()
}
\description{
Add a \verb{no data} (mask) color to legend
}
\references{
\url{https://stackoverflow.com/a/42380983}
}
