% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{agreement_frequency}
\alias{agreement_frequency}
\title{Frequency of elements in \code{x} with the same sign as \code{x_central}}
\usage{
agreement_frequency(x, x_central = NULL)
}
\description{
Frequency of elements in \code{x} with the same sign as \code{x_central}
}
\examples{
x <- seq(-5, 10)
agreement_frequency(x)
agreement_frequency(x, x_central = -10)
agreement_frequency(x, x_central = 0)

x <- matrix(rnorm(100), nrow = 20)
agreement_frequency(x)
agreement_frequency(x, x_central = -10)
agreement_frequency(x, x_central = 0)
agreement_frequency(x, x_central = c(rep(1, 10), rep(-1, 10)))

}
