% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__PredictionRR.R
\name{calc_MIRRP_from_dists}
\alias{calc_MIRRP_from_dists}
\title{Identify most important \code{R&R} predictor (\code{MIRRP})}
\usage{
calc_MIRRP_from_dists(x, tol = sqrt(.Machine[["double.eps"]]), seed = NULL)
}
\description{
We define the \code{MIRRP} as the predictor that is
associated with the largest, positive distance.
There is no \code{MIRRP} if maximum distance value is shared by
at least 50\% of predictors. Otherwise, ties are resolved at
random.
}
\examples{
x <- rbind(
  1:5,
  5:1,
  rep(4, 5),
  c(4, 4, 1, 1, 1),
  rep(-1, 5)
)
newRR3::calc_MIRRP_from_dists(x, seed = 123) # c(5, 1, NA, 1, NA)

}
