% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__PredictionRR.R
\name{calc_nearest_column}
\alias{calc_nearest_column}
\title{Identify column index of nearest value by rows}
\usage{
calc_nearest_column(x, mat, tol = sqrt(.Machine[["double.eps"]]), seed = NULL)
}
\description{
Identify column index of nearest value by rows
}
\examples{
loc <- c(1.5, 2, 4, 2)
mat <- rbind(
  1:5,
  5:1,
  rep(4, 5),
  c(4, 4, 2, 1, 1)
)
newRR3::calc_nearest_column(loc, mat) # c(1 or 2, 4, any of 1:5, 3)

}
