% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__PredictionRR.R
\name{calc_RF_response_from_probs}
\alias{calc_RF_response_from_probs}
\title{Calculate response level from class probabilities}
\usage{
calc_RF_response_from_probs(
  x,
  cutoffs_probs = rep(1/ncol(x), ncol(x)),
  seed = NULL
)
}
\arguments{
\item{x}{A two-dimensional numerical object. Rows represent sites and
columns represent class probabilities.}

\item{cutoffs_probs}{A numerical vector of length equal to number of columns
of \code{x}.}

\item{seed}{An integer value. Set the random number generator seed.}
}
\value{
An integer vector of length equal to the number of rows of \code{x}
representing the class (column of \code{x}) with the largest value, i.e.,
the most likely category.
}
\description{
From \code{randomForest:::predict.randomForest()} -> \code{classForest()}
\url{https://github.com/cran/randomForest/blob/6cedb5fbf26ca857afda687f42c5ba535d61eaf0/src/rf.c#L567}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  Aggregated prediction is the class with the maximum votes/cutoff
  crit = (countts[j + n * *nclass] / *ntree) / cutoff[j];
  ...
  Break ties at random:
}\if{html}{\out{</div>}}
}
\examples{
x <- rbind(
  c(0.1, 0.4, 0.4, 0.1),
  c(0.7, 0.1, 0.1, 0.1),
  c(NA, 0.1, 0.1, 0.1)
)
calc_RF_response_from_probs(x) ## expected outcome: c(2 or 3, 1, NA)

}
