% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__PredictionRR.R
\name{calc_RR_index}
\alias{calc_RR_index}
\title{Calculate weighted sum across class probabilities}
\usage{
calc_RR_index(x, weights = c(-1, -1/3, 1/3, 1))
}
\arguments{
\item{x}{A two-dimensional numerical object. Rows represent sites and
columns represent class probabilities.}

\item{weights}{A numerical vector of length equal to number of columns
of \code{x}.}
}
\value{
A numeric vector of length equal to the number of rows of \code{x}
representing the weighted sum of the class probabilities.
}
\description{
Default weights applied to resilience and resistance categories
low, medium-low, medium, and high (the last including medium-high) results
in an index varying from -1 (low probabilities dominate) to
+1 (high probabilities dominate) with assumptions of distances among
classes being constant and equal.
}
\examples{
x <- rbind(
  c(0.1, 0.4, 0.4, 0.1),
  c(0.7, 0.1, 0.1, 0.1),
  c(NA, 0.1, 0.1, 0.1),
  c(1, 0, 0, 0),
  c(0, 1, 0, 0),
  c(0, 0, 0, 1),
  c(0.5, 0, 0, 0.5),
  c(0, 0.5, 0.5, 0)
)
calc_RR_index(x) ## expected outcome: c(0, -0.6, NA, -1, -1 / 3, 1, 0, 0)

p <- seq(0, 1, by = 0.01)
xg <- expand.grid(L = p, ML = p, M = p, H = p)
ids1 <- abs(rowSums(xg) - 1) < sqrt(.Machine[["double.eps"]])
xg <- as.matrix(xg[ids1, , drop = FALSE])
ri <- calc_RR_index(xg)

if (requireNamespace("graphics") && requireNamespace("stats")) {
  par_prev <- graphics::par(
    mfrow = c(2L, 2L),
    mar = c(2.5, 2.5, 0.5, 0.5),
    mgp = c(1, 0, 0),
    tcl = 0.3
  )
  for (k in 1:4) {
    x <- xg[, k]
    graphics::plot(
      x,
      ri,
      pch = 16,
      col = "darkgray",
      xlab = paste0("Probability (", colnames(xg)[k], ")"),
      ylab = "Continuous index"
    )
    graphics::lines(
      p,
      stats::predict(stats::lm(ri ~ x), newdata = data.frame(x = p)),
      col = "red",
      lwd = 2
    )
  }
  graphics::par(par_prev)
}

}
