% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colors_NA_Chambers2023}
\alias{colors_NA_Chambers2023}
\title{Color used for \verb{no data} in Chambers et al. 2023}
\usage{
colors_NA_Chambers2023()
}
\description{
Color used for \verb{no data} in Chambers et al. 2023
}
