% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colors_RR_Chambers2023}
\alias{colors_RR_Chambers2023}
\title{Colors used for \code{R&R} levels in Chambers et al. 2023}
\usage{
colors_RR_Chambers2023(rr_levels = c("L", "ML", "M", "H+MH"))
}
\description{
Colors used for \code{R&R} levels in Chambers et al. 2023
}
