% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_flatbinned}
\alias{colorscale_flatbinned}
\title{Binned color scale where part of the scale is flat (one consistent color)}
\usage{
colorscale_flatbinned(
  name,
  breaks,
  threshold,
  limits = range(breaks),
  aesthetics = "fill",
  color_flat = "antiquewhite",
  position_flat = 1,
  viridis_pal_option = "C",
  reverse = FALSE,
  begin_hue = 0,
  end_hue = 1,
  ...
)
}
\arguments{
\item{breaks}{Integer vector with breaks including lower and upper limit.}
}
\description{
Binned color scale where part of the scale is flat (one consistent color)
}
