% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_flatcontinuous}
\alias{colorscale_flatcontinuous}
\title{Color scale to display a combination of fixed/flat color and continuous scale}
\usage{
colorscale_flatcontinuous(
  name,
  threshold,
  limits,
  aesthetics = "fill",
  color_flat = "antiquewhite",
  position_flat = 1,
  viridis_pal_option = "C",
  reverse = FALSE,
  begin_hue = 0,
  end_hue = 1,
  ...
)
}
\arguments{
\item{position_flat}{1: values between \code{limits[1]} and \code{threshold}: fixed \code{color_flat}
values between \code{threshold} and \code{limits[2]}: \code{viridis} "plasma"
2: values between \code{limits[1]} and threshold: \code{viridis} "plasma"
values between \code{threshold} and \code{limits[2]}: fixed \code{color_flat}}
}
\description{
Color scale to display a combination of fixed/flat color and continuous scale
}
