% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_robustdelta_by_ref}
\alias{colorscale_robustdelta_by_ref}
\title{Color scale to display robust deltas grouped by original level with \code{ggplot2}}
\usage{
colorscale_robustdelta_by_ref(name, levels, aesthetics = "fill", ...)
}
\description{
Color scale to display robust deltas grouped by original level with \code{ggplot2}
}
