% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_robustsimpledelta}
\alias{colorscale_robustsimpledelta}
\title{Color scale to display robust simple deltas with \code{ggplot2}}
\usage{
colorscale_robustsimpledelta(name, levels, aesthetics = "fill", ...)
}
\description{
Color scale to display robust simple deltas with \code{ggplot2}
}
