% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_rr}
\alias{colorscale_rr}
\title{Color scale to display \code{R&R} indicators}
\usage{
colorscale_rr(name, levels, aesthetics = "fill", ...)
}
\description{
Color scale to display \code{R&R} indicators
}
