% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{colorscale_rr2}
\alias{colorscale_rr2}
\title{Continuous approach to coloring RR indicators}
\usage{
colorscale_rr2(name, levels, aesthetics = "fill", ...)
}
\description{
Continuous approach to coloring RR indicators
}
