% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{create_crosshatching}
\alias{create_crosshatching}
\title{Create a polygon from raster that can be plotted as cross-hatching layer}
\usage{
create_crosshatching(
  xstars,
  x,
  rat_values,
  dx = units::set_units(4L, "km"),
  fun_crosshatch = function(x) ifelse(x >= 0.9, NA, 1L)
)
}
\description{
Create a polygon from raster that can be plotted as cross-hatching layer
}
