% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{create_tu_tag}
\alias{create_tu_tag}
\title{Create identifying tag, suitable for file names}
\usage{
create_tu_tag(x)
}
\arguments{
\item{x}{A character string or vector of character strings.}
}
\value{
A copy of \code{x} where multiple strings are collapsed by \code{"_"} and
all non alpha-numeric characters removed.
}
\description{
Create identifying tag, suitable for file names
}
\examples{
create_tu_tag("test12a")
create_tu_tag(c("test12a", "test12b", "test13a"))
create_tu_tag("_test12a.")
create_tu_tag(c("_test12a.", "/test12b?", "-test13a|"))

}
