% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_EPAecoregionL3_polygon}
\alias{download_EPAecoregionL3_polygon}
\title{Download EPA ecoregions L3 polygon}
\usage{
download_EPAecoregionL3_polygon(path, method_to_import = "download")
}
\description{
Download EPA ecoregions L3 polygon
}
\references{
\url{https://www.epa.gov/eco-research/ecoregions-north-america}
}
