% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_ESRIfederallands_polgyon}
\alias{download_ESRIfederallands_polgyon}
\title{USA Federal Lands by ESRI}
\usage{
download_ESRIfederallands_polgyon(path, method_to_import = "download")
}
\description{
USA Federal Lands by ESRI
}
\references{
\url{https://www.arcgis.com/home/item.html?id=5e92f2e0930848faa40480bcb4fdc44e#overview}
as referenced by \url{https://www.sciencebase.gov/catalog/item/4fc50e63e4b00e9c12d8c322}

Last download on 2022-Aug-16
}
