% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_from_ScienceBase}
\alias{download_from_ScienceBase}
\title{Download all data attached to a Science Base Data Release}
\usage{
download_from_ScienceBase(
  doi,
  fnames_downloaded_data,
  fnames_downloadraw,
  dir_unzip = NULL,
  dir_downloadraw = unique(dirname(fnames_downloadraw))
)
}
\description{
Download all data attached to a Science Base Data Release
}
