% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_NEadmin1_polygon}
\alias{download_NEadmin1_polygon}
\title{Download Natural Earth's North American political boundaries "admin1"}
\usage{
download_NEadmin1_polygon(path, method_to_import = "download")
}
\description{
Download Natural Earth's North American political boundaries "admin1"
}
\references{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
