% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_sagebiome_polygon}
\alias{download_sagebiome_polygon}
\title{Download the "Sagebrush Biome" polygon}
\usage{
download_sagebiome_polygon(path, method_to_import = "download")
}
\description{
Download the "Sagebrush Biome" polygon
}
\references{
Jeffries, M.I., and Finn, S.P., 2019,
The Sagebrush Biome Range Extent, as Derived from Classified Landsat Imagery:
U.S. Geological Survey data release, \url{https://doi.org/10.5066/P950H8HS}
}
