% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_scd1sei_data}
\alias{download_scd1sei_data}
\title{Download 30-m raster of Sagebrush Ecological Integrity (\code{"SEI"})
by the Sagebrush Conservation Design}
\usage{
download_scd1sei_data(path, method_to_import = "download")
}
\description{
Download 30-m raster of Sagebrush Ecological Integrity (\code{"SEI"})
by the Sagebrush Conservation Design
}
\references{
Doherty, K., Theobald, D.M., Holdrege, M.C., Wiechman, L.A., and
Bradford, J.B., 2022, Biome-wide sagebrush core habitat and growth areas
estimated from a threat-based conservation design:
U.S. Geological Survey data release, \url{https://doi.org/10.5066/P94Y5CDV}
}
