% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{download_SMAfederallands_polgyon}
\alias{download_SMAfederallands_polgyon}
\title{Download BLM National Surface Management Agency Area Polygons -
National Geospatial Data Asset (\code{NGDA})}
\usage{
download_SMAfederallands_polgyon(path, method_to_import = "download")
}
\description{
Download BLM National Surface Management Agency Area Polygons -
National Geospatial Data Asset (\code{NGDA})
}
\references{
\url{https://catalog.data.gov/dataset/blm-national-surface-management-agency-area-polygons-national-geospatial-data-asset-ngda},
referenced by \url{https://www.sciencebase.gov/catalog/item/59b83c14e4b08b1644df5f6e}
}
