% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{fun_acrossgeo_categorical}
\alias{fun_acrossgeo_categorical}
\title{Counts of an integer or categorical variable in a data frame}
\usage{
fun_acrossgeo_categorical(x, var, var_count = "count.x", levels = NULL, ...)
}
\arguments{
\item{x}{A two dimensional object with columns names corresponding to
\code{var} and \code{var_count}.}

\item{var}{A character string identifying the variable in \code{x} for which
to sum up counts.}

\item{var_count}{A character string identifying the column in \code{x} that
provides the counts for \code{var}.}

\item{levels}{A vector of character strings with (possible) levels of
\code{x[[var]]} for which a count is returned.}

\item{...}{Silently ignored.}
}
\value{
A named numeric vector of length equal to the length of \code{levels}
(or unique elements of \code{x[[var]]} if \code{levels} is \code{NULL})
plus 1 (the count of \code{NA}).
}
\description{
Counts of an integer or categorical variable in a data frame
}
\examples{
x <- cbind(
  x = c(rep(1, 5), 4, rep(2, 4), NA),
  count.x = c(rep(4, 5), rep(1, 3), rep(0, 2), 1)
)
fun_acrossgeo_categorical(x, var = "x")
fun_acrossgeo_categorical(x, var = "x", levels = 0:5)

x <- data.frame(
  x = c(rep("a", 5), "d", rep("b", 4), NA),
  count.x = c(rep(4, 5), rep(1, 3), rep(0, 2), 1)
)
fun_acrossgeo_categorical(x, var = "x")
fun_acrossgeo_categorical(x, var = "x", levels = letters[1:5])
fun_acrossgeo_categorical(x, var = "x", levels = c("b", "e", "d"))

}
\seealso{
\code{\link[=fun_acrossgeo_numeric]{fun_acrossgeo_numeric()}}
}
