% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{fun_acrossgeo_numeric}
\alias{fun_acrossgeo_numeric}
\title{Weighted quantiles of a variable in a data frame}
\usage{
fun_acrossgeo_numeric(
  x,
  var,
  var_count = "count.x",
  probs = c(0.05, 0.5, 0.95),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{A two dimensional object with columns names corresponding to
\code{var} and \code{var_count}.}

\item{var}{A character string identifying the variable in \code{x} for which
to take quantiles.}

\item{var_count}{A character string identifying the column in \code{x} that
provides the frequency weights for \code{var}.}

\item{probs}{A numeric vector of probabilities with values in \verb{[0, 1]}.}

\item{na.rm}{A logical value.}

\item{...}{Silently ignored.}
}
\value{
A named numeric vector of length equal to the length of \code{probs}.
}
\description{
Weighted quantiles of a variable in a data frame
}
\examples{
x <- cbind(
  x = c(1:10, NA),
  count.x = c(rep(4, 5), rep(1, 3), rep(0, 2), 1)
)
fun_acrossgeo_numeric(x, var = "x")
fun_acrossgeo_numeric(x, var = "x", na.rm = FALSE)
fun_acrossgeo_numeric(x[1:8, ], var = "x")

}
\seealso{
\code{\link[Hmisc:wtd.stats]{Hmisc::wtd.quantile()}}, \code{\link[=fun_acrossgeo_categorical]{fun_acrossgeo_categorical()}}
}
