% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{get_project_description}
\alias{get_project_description}
\title{List project description / metadata}
\usage{
get_project_description(
  include_variables = FALSE,
  include_scenarios = FALSE,
  include_deltas = FALSE,
  include_summaries = FALSE,
  dir_data = "data-raw",
  tag_prj = "newRR_wall2wall_future",
  simexps = c("hist", "rcp45", "rcp85"),
  varsets_tags = c(preds = "Preds19", novelty = "Preds19-Novelty", rr = "RR", rfcertainty
    = "RR-RFcertainty", mirrp = "MIRRP"),
  versions = list(version_inputs = "20220314", version_simexp = "20220408",
    version_simdata = "20221001", version_rf = "v20220324rf"),
  tag_sims = "SOILWAT2_simulations",
  isClimatology = TRUE,
  tag_scen1 = "ambient_1980-2020"
)
}
\arguments{
\item{include_variables}{A logical value. Include description about
R&R categories and R&R predictor variables. Additional information is loaded
from disk using \code{dir_data}.}

\item{include_scenarios}{A logical value. Include description about
climate scenarios represented by \code{rSFSW2} simulation experiments.
Additional information is loaded from disk using \code{dir_data}.}

\item{include_deltas}{A logical value. Include descriptions and functions
to determine changes between a scenario and a reference.}

\item{include_summaries}{A logical value. Include descriptions and
functions to summarize values across space and across \code{GCMs}.}

\item{dir_data}{A character string. The path to the local \code{data} folder,
required only if \code{include_variables} and/or \code{include_scenarios}.}
}
\value{
A named list.
}
\description{
List project description / metadata
}
\examples{
meta <- newRR3::get_project_description(
  include_variables = TRUE,
  include_scenarios = TRUE,
  include_deltas = TRUE,
  include_summaries = TRUE,
  dir_data = system.file("extdata", "data-raw", package = "newRR3"),
  tag_prj = "newRR3",
  tag_sims = ""
)

}
