% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__Novelty.R
\name{get_training_rfworkflows}
\alias{get_training_rfworkflows}
\title{Load and extract training data from fitted models for R&R}
\usage{
get_training_rfworkflows(fname_training, fnames_mrrrf)
}
\arguments{
\item{fname_training}{File name with path for the
extracted training data (stored as \code{"rds"} file).}

\item{fnames_mrrrf}{A named vector of length two with file names to the
fitted workflow objects for \code{"Resilience"} and \code{"Resistance"} stored as
\code{"rds"} files on disk. Only used if \code{fname_training} doesn't already exist.}
}
\value{
A data frame with the training data.
}
\description{
Load training data if already extracted from fitted models for R&R;
otherwise, extract training data and write to disk.
}
