% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{ggplot2_clean_theme}
\alias{ggplot2_clean_theme}
\title{Clean theme for \code{ggplot2}
Based on \code{egg::theme_article()}}
\usage{
ggplot2_clean_theme()
}
\description{
Clean theme for \code{ggplot2}
Based on \code{egg::theme_article()}
}
