% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_EPAecoregionL3_polygon}
\alias{import_EPAecoregionL3_polygon}
\title{Import EPA ecoregions L3 polygon}
\usage{
import_EPAecoregionL3_polygon(dir_dataraw = file.path("..", "data-raw"), ...)
}
\description{
Import EPA ecoregions L3 polygon
}
