% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_ESRIfederallands_polgyon}
\alias{import_ESRIfederallands_polgyon}
\title{Import USA Federal Lands by ESRI}
\usage{
import_ESRIfederallands_polgyon(dir_dataraw = file.path("..", "data-raw"), ...)
}
\description{
Import USA Federal Lands by ESRI
}
