% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_NEadmin1_polygon}
\alias{import_NEadmin1_polygon}
\title{Import Natural Earth's North American political boundaries of states
and provinces}
\usage{
import_NEadmin1_polygon(dir_dataraw = file.path("..", "data-raw"), ...)
}
\description{
Import Natural Earth's North American political boundaries of states
and provinces
}
