% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_predictors_rsw2metrics}
\alias{import_predictors_rsw2metrics}
\title{Import \code{rSW2metrics} extraction object with predictors for R&R models}
\usage{
import_predictors_rsw2metrics(
  dir_dataraw = file.path("..", "data-raw"),
  dir_extractmetrics = file.path("..", "..", "2_CalculateDerivedMetrics"),
  metric_name = "RR2022predictors",
  list_simexps = NULL,
  version_simdata = NULL,
  tag_simexp0 = NULL,
  method_to_import = c("symlink", "copy")
)
}
\description{
Import \code{rSW2metrics} extraction object with predictors for R&R models
}
