% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_predictors_variablenames}
\alias{import_predictors_variablenames}
\title{Import predictor variable names for R&R models (Chambers et al. 2023)}
\usage{
import_predictors_variablenames(
  dir_dataraw = file.path("..", "data-raw"),
  dir_plotsim = file.path("..", "..", "..", "Prod043b_newRR_PlotSimulations"),
  method_to_import = c("symlink", "copy")
)
}
\description{
Import predictor variable names for R&R models (Chambers et al. 2023)
}
