% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_predictrr_rfworkflows}
\alias{import_predictrr_rfworkflows}
\title{Import workflows to predict R&R (Chambers et al. 2023)}
\usage{
import_predictrr_rfworkflows(
  dir_dataraw = file.path("..", "data-raw"),
  dir_plotsim = file.path("..", "..", "..", "Prod043b_newRR_PlotSimulations"),
  version_rf = NULL,
  method_to_import = c("symlink", "copy")
)
}
\description{
Import workflows to predict R&R (Chambers et al. 2023)
}
