% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_sagebiome_polygon}
\alias{import_sagebiome_polygon}
\title{Import the "Sagebrush Biome" polygon}
\usage{
import_sagebiome_polygon(dir_dataraw = file.path("..", "data-raw"), ...)
}
\description{
Import the "Sagebrush Biome" polygon
}
