% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_scd1sei_raster}
\alias{import_scd1sei_raster}
\title{Import 30-m raster of Sagebrush Ecological Integrity (\code{SEI})
by the Sagebrush Conservation Design}
\usage{
import_scd1sei_raster(
  dir_dataraw = file.path("..", "data-raw"),
  years = NA,
  ...
)
}
\description{
Import 30-m raster of Sagebrush Ecological Integrity (\code{SEI})
by the Sagebrush Conservation Design
}
