% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_scenarios_rsfsw2}
\alias{import_scenarios_rsfsw2}
\title{Import climate scenarios represented by \code{rSFSW2} experiments}
\usage{
import_scenarios_rsfsw2(
  dir_dataraw = file.path("..", "data-raw"),
  dir_simexps = file.path("..", "..", "1_SOILWAT2_Simulations"),
  tag_simexp = NULL,
  method_to_import = c("symlink", "copy")
)
}
\description{
Import climate scenarios represented by \code{rSFSW2} experiments
}
