% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_SMAfederallands_polgyon}
\alias{import_SMAfederallands_polgyon}
\title{Import USA Federal Lands by \code{SMA}}
\usage{
import_SMAfederallands_polgyon(dir_dataraw = file.path("..", "data-raw"), ...)
}
\description{
Import USA Federal Lands by \code{SMA}
}
