% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ExternalDataImport.R
\name{import_to_project}
\alias{import_to_project}
\title{Import files into project by symbolic linking or by copying}
\usage{
import_to_project(
  files_imported,
  files_external,
  method = c("symlink", "copy"),
  add_nparents = 0L
)
}
\arguments{
\item{files_imported}{File path(s) relative to current working directory.}

\item{files_external}{File path(s) relative to current working directory.}

\item{method}{A character string, either \code{"symlink"} or \code{"copy"}.}

\item{add_nparents}{An integer value. The level of sub-folders where
the \code{"symlink"}s will be created relative to current working directory.}
}
\description{
Import files into project by symbolic linking or by copying
}
