% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{inset_densitycountplot}
\alias{inset_densitycountplot}
\title{Create minimal plot of weighted (count * density)}
\usage{
inset_densitycountplot(
  x,
  limits = c(NA, NA),
  x_binned = FALSE,
  weight = 1,
  consistency = NULL,
  consistency_lvls = NULL,
  add_vertical0 = TRUE,
  show.legend = FALSE
)
}
\description{
Create minimal plot of weighted (count * density)
}
