% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{integerize_rat_header}
\alias{integerize_rat_header}
\title{Make sure \code{RAT} fields "value" and "count" are integers}
\usage{
integerize_rat_header(x, headers = c("^value$", "^count"))
}
\arguments{
\item{x}{A data frame used as raster attribute value table}
}
\value{
A data frame corresponding to \code{x} where
columns matching \code{"value"} and \code{"count"} are converted to integers.
}
\description{
Make sure \code{RAT} fields "value" and "count" are integers
}
\examples{
n <- 3L
x <- data.frame(
  value = as.numeric(seq_len(n)),
  count = sample(10, size = n),
  var1 = runif(n = n),
  var2 = letters[seq_len(n)]
)

xi <- integerize_rat_header(x)

}
