% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__Novelty.R
\name{is_novel}
\alias{is_novel}
\alias{is_novel_lvls}
\title{Determine if \code{NT2} or distance-ratio of \code{AOA} represents novel conditions}
\usage{
is_novel(x)

is_novel_lvls(x)
}
\arguments{
\item{x}{A numerical vector with \code{NT2} or distance-ratios of \code{AOA}.}
}
\value{
A logical vector of the same length as \code{x} for \code{is_novel()};
an integer vector of the same length as \code{x} for \code{is_novel_lvls()} where
a value of \code{1} indicates \code{"novel"} and a value of \code{2} indicates \code{"notnovel"}.
}
\description{
Determine if \code{NT2} or distance-ratio of \code{AOA} represents novel conditions
}
\examples{
xt <- seq(0, 2, by = 0.5)
cbind(x = xt, lgl = is_novel(xt), lvls = is_novel_lvls(xt))

}
