% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{locate_tabulated_files}
\alias{locate_tabulated_files}
\title{Locate tabulated \code{csv}-spreadsheet files}
\usage{
locate_tabulated_files(filenames, simslices, extent, variables)
}
\description{
Locate tabulated \code{csv}-spreadsheet files
}
