% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{make_acrmod_labels}
\alias{make_acrmod_labels}
\title{Create nice labels from across-model tags}
\usage{
make_acrmod_labels(x)
}
\description{
Create nice labels from across-model tags
}
\examples{
make_acrmod_labels(c("RCP45_2029-2064", "RCP45_2064-2099"))
make_acrmod_labels("historical_1950-2005")

}
