% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{make_convolution}
\alias{make_convolution}
\title{Description of convolution of a categorical variable with itself}
\usage{
make_convolution(levels)
}
\description{
Description of convolution of a categorical variable with itself
}
