% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{make_convolution_2way}
\alias{make_convolution_2way}
\title{Description of convolution of two categorical variables}
\usage{
make_convolution_2way(
  levels1,
  levels2,
  sep = "_from_",
  signs = NULL,
  convolve_method = c("to-from", "from-to")
)
}
\description{
Description of convolution of two categorical variables
}
