% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{make_geotiff}
\alias{make_geotiff}
\title{Make a (\code{COG}) \code{GeoTIFF}}
\arguments{
\item{rast_template}{A file name or a \code{terra} \code{SpatRaster} object.}

\item{rat}{A \code{data.frame} object. The raster (value) attribute table
where the first column is \code{"Value"} that are matched against the values
of \code{rast_template}.}

\item{fname}{A file name. The file name of the resulting \code{GeoTIFF}.}

\item{names_rat}{The names assigned to the layers of the \code{SpatRaster}
(if not missing).}

\item{add_esri_vat}{A logical value. If \code{TRUE}, then
write an \code{ESRI}-formatted value attribute table
(in addition to the \code{GDAL}-formatted raster attribute table).}

\item{vat_rename_cns}{List to rename attributes for the value attribute table
(\code{ESRI} attribute names are restricted to 10 or fewer characters).}

\item{overwrite}{A logical value. Should an existing \code{fname} be overwritten?}

\item{filetype}{Passed to \link[terra:writeRaster]{terra::writeRaster}, e.g., \code{"GTiff"}, \code{"COG"}}

\item{gdal_write_options}{Passed to \link[terra:writeRaster]{terra::writeRaster}.}
}
\description{
Make a (\code{COG}) \code{GeoTIFF}
}
