% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{make_geotiff_catalyzed}
\alias{make_geotiff_catalyzed}
\title{Create a \code{GeoTIFF} from template \code{GeoTIFF} using values from a spreadsheet}
\usage{
make_geotiff_catalyzed(
  rast_template,
  rat,
  fname,
  names_rat,
  rat2 = NULL,
  add_esri_vat = TRUE,
  vat_rename_cns = NULL,
  overwrite = FALSE,
  filetype = "GTiff",
  gdal_write_options = c("COMPRESS=LZW", "TFW=YES", "TILED=YES")
)
}
\arguments{
\item{rast_template}{A file name or a \code{terra} \code{SpatRaster} object.}

\item{rat}{A \code{data.frame} object. The raster (value) attribute table
where the first column is \code{"Value"} that are matched against the values
of \code{rast_template}.}

\item{fname}{A file name. The file name of the resulting \code{GeoTIFF}.}

\item{names_rat}{The names assigned to the layers of the \code{SpatRaster}
(if not missing).}

\item{rat2}{A \code{data.frame} object. A secondary raster (value) attribute table
that can be used to describe category levels
(useful if the content from \code{rat} that is catalyzed into
the resulting \code{GeoTIFF} are categorical).}

\item{add_esri_vat}{A logical value. If \code{TRUE}, then
write an \code{ESRI}-formatted value attribute table
(in addition to the \code{GDAL}-formatted raster attribute table).}

\item{vat_rename_cns}{List to rename attributes for the value attribute table
(\code{ESRI} attribute names are restricted to 10 or fewer characters).}

\item{overwrite}{A logical value. Should an existing \code{fname} be overwritten?}

\item{filetype}{Passed to \code{\link[terra:writeRaster]{terra::writeRaster}}, e.g., \code{"GTiff"}, \code{"COG"}}

\item{gdal_write_options}{Passed to \code{\link[terra:writeRaster]{terra::writeRaster}}.}
}
\description{
Create a \code{GeoTIFF} from template \code{GeoTIFF} using values from a spreadsheet
}
\seealso{
\code{\link[=make_geotiff_with_rat]{make_geotiff_with_rat()}}
}
