% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{make_geotiff_with_rat}
\alias{make_geotiff_with_rat}
\title{Create a \code{GeoTIFF} from template \code{GeoTIFF} after adding spreadsheet as \code{RAT}}
\usage{
make_geotiff_with_rat(
  rast_template,
  rat,
  fname,
  names_rat,
  add_esri_vat = TRUE,
  vat_rename_cns = NULL,
  overwrite = FALSE,
  gdal_write_options = c("COMPRESS=LZW", "TFW=YES", "TILED=YES")
)
}
\arguments{
\item{rast_template}{A file name or a \code{terra} \code{SpatRaster} object.
The template raster passed to \link[terra:rast]{terra::rast}.}

\item{rat}{A \code{data.frame} object. The raster (value) attribute table
where the first column is \code{"Value"} that are matched against the values
of \code{rast_template}.}

\item{fname}{A file name. The file name of the resulting \code{GeoTIFF}.}

\item{names_rat}{The names assigned to the layers of the \code{SpatRaster}
(if not missing).}

\item{add_esri_vat}{A logical value. If \code{TRUE}, then
write an \code{ESRI}-formatted value attribute table
(in addition to the \code{GDAL}-formatted raster attribute table).}

\item{vat_rename_cns}{List to rename attributes for the value attribute table
(\code{ESRI} attribute names are restricted to 10 or fewer characters).}

\item{overwrite}{A logical value. Should an existing \code{fname} be overwritten?}

\item{gdal_write_options}{Passed to \link[terra:writeRaster]{terra::writeRaster}.}
}
\description{
Create a \code{GeoTIFF} from template \code{GeoTIFF} after adding spreadsheet as \code{RAT}
}
\seealso{
\code{\link[=make_geotiff_catalyzed]{make_geotiff_catalyzed()}}
}
