% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRR3-package.R
\docType{package}
\name{newRR3}
\alias{newRR3}
\alias{newRR3-package}
\title{\pkg{newRR3}: Analysis Framework for \code{newRR} Project.}
\description{
Analysis Framework for newRR Project.
}
\section{LICENSE}{

Copyright (C) 2023 by
newRR3 maintainer

Permission is hereby granted, free of charge, to any person obtaining a
copy of this software and associated documentation files (the "Software"),
to deal in the Software without restriction, including without limitation
the rights to use, copy, modify, merge, publish, distribute, sublicense,
and/or sell copies of the Software, and to permit persons to whom the
Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included
in all copies or substantial portions of the Software.
}

\section{DISCLAIMER}{

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}

\author{
\strong{Maintainer}: Daniel Schlaepfer \email{daniel.schlaepfer@alumni.ethz.ch} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
