% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{output_names}
\alias{output_names}
\alias{name_output_folder}
\alias{name1_output_ratcsv}
\alias{name2_output_ratcsv}
\title{Output folder and file names}
\usage{
name_output_folder(experiment_time, outputset, valuetype, variableset)

name1_output_ratcsv(variableset, modelconditions, methoddescription)

name2_output_ratcsv(variableset, modelconditions, valuetype, outputset)
}
\description{
Output folder and file names
}
\section{Functions}{
\itemize{
\item \code{name_output_folder()}: Output sub-directory name

\item \code{name1_output_ratcsv()}: Three part output file name

\item \code{name2_output_ratcsv()}: Output file name with value type and output set

}}
\section{Name scheme}{

\itemize{
\item Name scheme for output folders is
\code{"[experiment]_[time]_[output-set-dir]_[value-type-dir]_[variable-set]"}
\item Name scheme for output files with raster attribute tables is
\code{"[variable-set]__[model-conditions](__[method-description(s)])+.csv"}

The scheme expands, for output from individual simulation runs, to
\code{"[variable-set]__[scenario-id]_[GCM]_[experiment]_[time]__[value-type-file]-[output-set-file].csv"}

and, for output from across-model summaries, to
\verb{[variable-set]__[experiment]_[time]__[value-type-file]-[output-set-file]-[acrmod-summary].csv"}
\item Name scheme for output files with across-geography summaries is
\verb{[variable-set]__[experiment]_[time]__[method-description]__[acrgeo-summary]__[variable-name].csv"}
}
}

\section{Name elements}{

The name scheme uses a fixed vocabulary with the following elements:
\itemize{
\item \code{variable-set} = \verb{\{Preds19, Preds19-Novelty, RR, RR-RFcertainty\}}
\item \code{variable-name} = \verb{\{[variable names]\}}
\item \code{scenario-id} = \code{{sc1..sc41}}
\item \code{GCM} = \verb{\{[global climate models]\}}
\item \code{experiment} = \verb{\{ambient, historical, RCP45, RCP85\}}
\item \code{time} = \verb{\{1980-2020-clim, 1950-2005-clim, 2029-2064-clim, 2064-2099-clim\}}
\item \code{output-set-dir} = \verb{\{IndRuns, AcrMod\}}
\item \code{output-set-file} = \verb{\{sim, acrmod\}}
\item \code{value-type-dir} = \verb{\{Values, Deltas\}}
\item \code{value-type-file} = \verb{\{value, delta, robust090delta, simpledelta, robust090simpledelta\}}
\item \code{acrmod-summary} = \verb{\{low, med, high, agree, robust090\}}
\item \code{acrgeo-summary} = \verb{\{argeo-[spatial units]\}}
}
}

