% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{panels_remove_inner_labels}
\alias{panels_remove_inner_labels}
\title{Remove inner labels from plot panels}
\usage{
panels_remove_inner_labels(
  list_panels,
  ks_panels_do_remove = NULL,
  remove_xlabs = NULL,
  remove_xticks = NULL,
  remove_ylabs = NULL,
  remove_yticks = NULL,
  remove_facetlabs = NULL,
  remove_inner_plotmargins = NULL,
  n_panels = NULL,
  byrow = TRUE
)
}
\arguments{
\item{ks_panels_do_remove}{An integer vector of panel numbers for which
inner labels actually should be removed. \code{NULL} (default) removes
inner labels from all panels.}

\item{remove_xlabs}{Remove x-axis label and x-axis tick labels.}

\item{remove_xticks}{Remove x-axis ticks.}
}
\description{
Remove inner labels from plot panels
}
