% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{plot_map}
\alias{plot_map}
\title{Create one map}
\usage{
plot_map(
  xstars,
  add_coords = TRUE,
  crs = sf::st_crs(xstars),
  bbox = sf::st_bbox(xstars),
  expand_bbox = TRUE,
  maintitle = NULL,
  subtitle = NULL,
  panel_tag = NULL,
  show_legend = TRUE,
  legend_xoffset = 0,
  legend_yoffset = 0,
  st_poly_hatched = NULL,
  st_geom_sb = NULL,
  st_geom_state = NULL,
  st_geom_eco = NULL
)
}
\description{
Create one map
}
