% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{plot_spatialtabulation_stackedbars}
\alias{plot_spatialtabulation_stackedbars}
\title{Create stacked bar plot for spatial tabulations}
\usage{
plot_spatialtabulation_stackedbars(data, var_y, colorscale)
}
\description{
Create stacked bar plot for spatial tabulations
}
