% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__DisplayItems.R
\name{plot_spatialtabulation_stackedbars3}
\alias{plot_spatialtabulation_stackedbars3}
\title{Create stacked bar plot for spatial tabulations, grouped for low, med, high}
\usage{
plot_spatialtabulation_stackedbars3(data, var_y, colorscale)
}
\description{
Create stacked bar plot for spatial tabulations, grouped for low, med, high
}
\references{
for variable-width stacked bars

\url{https://stackoverflow.com/questions/5948604/variable-width-bars-in-ggplot2-barplot-in-r}

\url{https://stackoverflow.com/questions/9556359/different-width-for-each-bar-using-geom-bar-with-position-fill-in-ggplot2}
}
