% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__PredictionRR.R
\name{predict_rr}
\alias{predict_rr}
\title{Predict resistance and resilience}
\usage{
predict_rr(
  x,
  var_predictors,
  rf_workflow_rst,
  rf_workflow_rsl,
  nrr,
  rr_levels = c("L", "ML", "M", "H+MH"),
  seed = 1234L
)
}
\arguments{
\item{x}{A data frame in wide format, i.e., columns represent predictor
variables and rows are treated as units.}

\item{var_predictors}{A vector of character strings. The names of the
predictor variables required by the workflows.}

\item{rf_workflow_rst}{A workflow object. A \code{"tidymodels"} workflow of a
\code{"randomForest"} model to predict resistance with \code{var_predictors} from \code{x}.}

\item{rf_workflow_rsl}{A workflow object. A \code{"tidymodels"} workflow of a
\code{"randomForest"} model to predict resilience with \code{var_predictors} from \code{x}.}

\item{nrr}{A vector of character strings. The names of the output variables.}

\item{rr_levels}{A vector of character strings. The names of
resistance and resilience categories (in increasing order).}

\item{seed}{An integer value. Random number generator seed; note that
predicted ties in most likely categories are broken at random.}
}
\value{
A data frame in wide format, i.e., columns represent predicted
variables (as identified by \code{nrr} plus any column copied from \code{x} not
identified as predictor by \code{var_predictors}) and rows are treated as units
(the same rows as \code{x}).
}
\description{
Predict resistance and resilience
}
