% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{prepare_df_for_stackedbars}
\alias{prepare_df_for_stackedbars}
\title{Read files and prepare data frame for stacked bar plots}
\usage{
prepare_df_for_stackedbars(filenames, simslices, extent, variables)
}
\description{
Read files and prepare data frame for stacked bar plots
}
