% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{process_raster}
\alias{process_raster}
\title{Load and downsample GeoTIFF with RAT as \code{terra} raster}
\usage{
process_raster(fname_geotiffs, downsample = 4L)
}
\description{
Load and downsample GeoTIFF with RAT as \code{terra} raster
}
