% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{robustify}
\alias{robustify}
\title{Combine categorical data with robust/unrobust classifier}
\usage{
robustify(x, xrobust, x_conv, robust_conv, robust_levels)
}
\arguments{
\item{x}{An integer vector. Factor values corresponding to
categories of \code{x_conv[["levels"]]}}
}
\description{
Combine categorical data with robust/unrobust classifier
}
\examples{
set.seed(3454)
N <- 10L

xlevels <- letters[1:4]
x_conv <- make_convolution(xlevels)
robust_levels <- c("nonrobust", "robust")
robust_conv <- make_2elems_convolution(
  levels1 = x_conv[["levels"]],
  levels2 = robust_levels,
  signs = x_conv[["sign"]]
)

x <- sample(
  x = seq_along(x_conv[["levels"]]),
  size = N,
  replace = TRUE
)
xrobust <- sample(
  x = seq_along(robust_levels) - 1L,
  size = N,
  replace = TRUE
)

rx <- robustify(x, xrobust, x_conv, robust_conv, robust_levels)

data.frame(
  x,
  xf = factor(
    x,
    levels = seq_along(x_conv[["levels"]]),
    labels = x_conv[["levels"]]
  ),
  xrobust,
  xrobustf = factor(
    1L + xrobust,
    levels = seq_along(robust_levels),
    labels = robust_levels
  ),
  rx,
  rxf = factor(
    rx,
    levels = seq_along(robust_conv[["levels"]]),
    labels = robust_conv[["levels"]]
  )
)

}
