% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{robustness}
\alias{robustness}
\title{Robustness is agreement larger than a frequency threshold}
\usage{
robustness(x, frq_robust)
}
\description{
Robustness is agreement larger than a frequency threshold
}
\examples{
x <- seq(-5, 10)
robustness(agreement_frequency(x), frq_robust = 0.9)

x <- matrix(rnorm(100), nrow = 20)
robustness(
  agreement_frequency(x, x_central = c(rep(1, 10), rep(-1, 10))),
  frq_robust = 0.7
)

}
