% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{select_function}
\alias{select_function}
\title{Select a function based on matching a name against listed options}
\usage{
select_function(
  x,
  options = list(list(vars = NULL, fun = NULL)),
  fun_default = function(x, y) x - y
)
}
\arguments{
\item{x}{One character string. The name for which a function is returned
either based on a match against an element of \code{options} or \code{fun_default}.}

\item{options}{A list or \code{NULL}. Elements of \code{options} consists of lists
with two named elements: \code{vars} (a vector of character strings against which
\code{x} is compared) and \code{fun} (a function to be returned if \code{x} matches an
element of \code{vars}; see requirements under \code{return}).}

\item{fun_default}{A function (see requirements under \code{return}).}
}
\value{
A function that takes two arguments \code{x} and \code{y}
(both vector of equal length) and
which returns a vector of equal length to \code{x} and \code{y}.
}
\description{
Select a function based on matching a name against listed options
}
\examples{
opts <- list(
  part1 = list(vars = c("a", "c", "d"), fun = function(x, y) paste(y, y)),
  part2 = list(vars = c("b", "e"), fun = function(x, y) paste(x, x))
)
deffun <- function(x, y) x + y

fun_d <- select_function("d", options = opts, deffun)
fun_e <- select_function("e", options = opts, deffun)
fun_f <- select_function("f", options = opts, deffun)

fun_d(0:2, 33:35)
fun_e(0:2, 33:35)
fun_f(0:2, 33:35)

}
