% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{shorten_variable_names}
\alias{shorten_variable_names}
\title{Make sure that all field names are shorter than \code{minlength} characters}
\usage{
shorten_variable_names(x, minlength = 10)
}
\description{
Make sure that all field names are shorter than \code{minlength} characters
}
\examples{
shorten_variable_names(c("short10", "verylongname"))
shorten_variable_names("mean_annual_temperature")

}
