% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{simplify_delta}
\alias{simplify_delta}
\title{Simplify convoluted categories (deltas of a category) to direction}
\usage{
simplify_delta(x, x_conv, simple_conv)
}
\arguments{
\item{x}{An integer vector representing convoluted levels.}

\item{x_conv}{A data frame describing convoluted \code{"levels"} of \code{x}
and their \code{"sign"}.}

\item{simple_conv}{A data frame describing simple \code{"levels"}
and their \code{"sign"}.}
}
\value{
An integer vector representing levels of
\code{simple_conv[["levels"]]} that were obtained from translating
\code{x} via their sign (i.e., direction) matched to the sign of levels
of \code{simple_conv}.
}
\description{
Simplify convoluted categories (deltas of a category) to direction
}
\examples{
Ncats <- 4L
n <- 10L
x <- sample(x = Ncats, size = n, replace = TRUE)
x_conv <- list(
  levels = letters[seq_len(Ncats)],
  sign = sign(seq_len(Ncats) - Ncats / 2L)
)
simple_conv <- list(
  levels = c("decrease", "stable", "increase"),
  sign = c(-1L, 0L, 1L)
)

xsimple <- simplify_delta(x, x_conv, simple_conv)

data.frame(
  x = x_conv[["levels"]][x],
  sign = x_conv[["sign"]][x],
  simple = simple_conv[["levels"]][xsimple]
)

}
\seealso{
\code{\link[=simplify_robustdelta]{simplify_robustdelta()}}
}
