% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{simplify_delta_by}
\alias{simplify_delta_by}
\title{Convert convoluted (change) categories to simplified change categories}
\usage{
simplify_delta_by(x, x_conv, sdo, by = c("from", "to"))
}
\description{
Convert convoluted (change) categories to simplified change categories
}
\examples{
n = 20L

# R & R categories and convolution
x_lvls <- c("L", "ML", "M", "H+MH")
x_conv <- make_convolution(x_lvls)

# Example data
x <- sample(x = nrow(x_conv), size = n, replace = TRUE)

# Simple decrease/stable/increase categories
dsi <- data.frame(
  levels = c("decrease", "stable", "increase"),
  sign = c(-1L, 0L, 1L)
)

# Target convolution of R&R categories and simple decreases/stable/increases
sdo_from <- make_2elems_convolution(
  levels1 = dsi[["levels"]],
  levels2 = x_lvls,
  signs = dsi[["sign"]],
  names = c("basic", "from")
)
sdo_to <- make_2elems_convolution(
  levels1 = dsi[["levels"]],
  levels2 = x_lvls,
  signs = dsi[["sign"]],
  names = c("basic", "to")
)

# Translation from convoluted R&R to simplified changes in R&R categories
xres <- simplify_delta_by(
  x = x,
  x_conv = x_conv,
  sdo = sdo_from,
  by = "from"
)

data.frame(
  x = x_conv[["levels"]][x],
  sign = x_conv[["sign"]][x],
  simple = sdo_from[["levels"]][xres]
)

xres <- simplify_delta_by(
  x = x,
  x_conv = x_conv,
  sdo = sdo_to,
  by = "to"
)

data.frame(
  x = x_conv[["levels"]][x],
  sign = x_conv[["sign"]][x],
  simple = sdo_to[["levels"]][xres]
)

}
