% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{simplify_robustdelta}
\alias{simplify_robustdelta}
\title{Simplify robust convoluted categories (deltas of a category) to direction}
\usage{
simplify_robustdelta(x, x_conv, simple_conv)
}
\arguments{
\item{x}{An integer vector representing robust convoluted levels.}

\item{x_conv}{A data frame describing \code{"robust"} convoluted \code{"levels"} of \code{x}
and their \code{"sign"}.}

\item{simple_conv}{A data frame describing \code{"robust"} simple \code{"levels"}
and their \code{"sign"}.}
}
\value{
An integer vector representing levels of
\code{simple_conv[["levels"]]} that were obtained from translating
\code{x} via their sign (i.e., direction) and robustness
matched to the sign and robustness of levels of \code{simple_conv}.
}
\description{
Simplify robust convoluted categories (deltas of a category) to direction
}
\examples{
Ncats <- 2L * 4L
n <- 20L
x <- sample(x = Ncats, size = n, replace = TRUE)

tmpr <- rep(c("nonrobust", "robust"), times = Ncats / 2L)
x_conv <- list(
  levels = paste0(tmpr, "_", rep(letters[seq_len(Ncats)], each = 2L)),
  robust = tmpr,
  sign = rep(sign(seq_len(Ncats) - Ncats / 2L), each = 2L)
)
tmpr <- rep(c("nonrobust", "robust"), times = 3L)
simple_conv <- list(
  levels = paste0(tmpr, "_", rep(c("decr", "stable", "incr"), each = 2L)),
  robust = tmpr,
  sign = rep(c(-1L, 0L, 1L), each = 2L)
)

xsimple <- simplify_robustdelta(x, x_conv, simple_conv)

data.frame(
  x = x_conv[["levels"]][x],
  sign = x_conv[["sign"]][x],
  simple = simple_conv[["levels"]][xsimple]
)

}
\seealso{
\code{\link[=simplify_delta]{simplify_delta()}}
}
