% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{simplify_variables_names}
\alias{simplify_variables_names}
\title{Remove suffices appended to derived variables}
\usage{
simplify_variables_names(
  x,
  tags_to_remove = c("delta", "robustdelta", "simpledelta", "acrmoddistr", "acrmodagree",
    "acrmodrobust")
)
}
\arguments{
\item{x}{A vector of character strings.}
}
\value{
A modified version of \code{x} representing base variable names.
}
\description{
Remove suffices appended to derived variables
}
\examples{
simplify_variables_names(
  c(
    "R_response",
    "R_prob.pred_H.MH_robustdelta_acrmoddistr",
    "R_prob.pred_H.MH_acrmoddistr_robustdelta"
  )
)

}
