% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ClimateModels.R
\name{sort_deltas_by_reference}
\alias{sort_deltas_by_reference}
\title{Sort RR (robust) deltas by reference category}
\usage{
sort_deltas_by_reference(x, ref_lvls, x_conv, simple_conv, is_robust)
}
\description{
Sort RR (robust) deltas by reference category
}
