% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__ProjectDescription-newRR3.R
\name{standardize_vtags}
\alias{standardize_vtags}
\title{Standardize variable group identifiers}
\usage{
standardize_vtags(x, tags)
}
\arguments{
\item{x}{A vector of character strings representing
variable group identifiers.}

\item{tags}{A named vector of character strings representing
all possible variable group identifiers where names represent standardized
names.}
}
\value{
A vector of character strings where each value of \code{x} is
translated in matching name of \code{tags}.
}
\description{
Standardize variable group identifiers
}
\examples{
tags <- c(first = "1st", second = "2nd", third = "3rd")
standardize_vtags(c("2nd", "5th", "3rd"), tags)

}
