% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__SpatialSummaries.R
\name{tabulate_acrgeo}
\alias{tabulate_acrgeo}
\title{Summarize across geographic space}
\usage{
tabulate_acrgeo(x, varname, is_var_categorical, ctus, fun, levels = NULL)
}
\arguments{
\item{x}{A two-dimensional object with columns \code{"value"}
(original spatial unit identifiers) and \code{varname}
(variable to be summarized); rows represent original spatial units.}

\item{varname}{A character string naming the variable to be summarized, i.e.,
a column name of \code{x}.}

\item{is_var_categorical}{A logical value to describe if \code{x[[var]]} contains
numerical or categorical data; see also \code{fun}.}

\item{ctus}{A list defining spatial units. Each element of \code{ctus} defines
one spatial unit across which \code{x} is summarized. Each of these elements
contains three named elements:
\itemize{
\item \code{"dft"}: a data frame where rows represent intersections between
spatial units to summarize over and the original spatial units
represented by rows of \code{x} and the two columns represent
\itemize{
\item \code{"suid"}: original spatial unit identifier linking to \code{x[["value"]]}
\item \code{"count"}: count of pixels within spatial intersection of
\code{x[["value"]]} and this element of \code{ctus}
}
\item \code{"ctus_orig"}: a vector of character strings naming the
original spatial units involved.
\item \code{"ctus"}: a character string representing the concatenated values of
\code{"ctus_orig"} and cleaned up to be used as part of a file name.
}}

\item{fun}{A function with arguments \code{x}, \code{var}, and \code{var_count}.
If \code{is_var_categorical} is \code{TRUE}, then \code{fun} should return the sums of
\code{x[[var_count]]} for each distinct value of \code{x[[var]]} and of \code{NA},
see \code{\link[=fun_acrossgeo_categorical]{fun_acrossgeo_categorical()}} for details.
If \code{is_var_categorical} is \code{TRUE}, then \code{fun} should return
quantiles of \code{x[[var]]} weighted by \code{x[[var_count]]},
see \code{\link[=fun_acrossgeo_numeric]{fun_acrossgeo_numeric()}} for details.}

\item{levels}{An optional vector of the unique values that
\code{x[[var]]} might have taken.
Passed to \code{fun} if \code{is_var_categorical} is \code{TRUE}.}
}
\value{
A data frame where rows represent spatial summaries for
each element of \code{ctus}. Columns include \code{"tag"} (the value of \code{"ctus"}),
columns for each element of \code{"ctus_orig"}, and columns of the
summarized values of \code{x} (corresponding to the output of \code{fun}).
}
\description{
Summarize across geographic space
}
\examples{
xn <- data.frame(value = 1:3, var = c(0.5, 0.1, 2))
xc1 <- data.frame(value = 1:3, var = c(2, 1, 4))
xc2 <- data.frame(value = 1:3, var = c("a", "c", "d"))

ctus <- list(
  list(
    dft = data.frame(suid = c(1, 3), count = c(100, 0)),
    ctus = "a1_b3",
    ctus_orig = c(polyA = "a1", polyB = "b3")
  ),
  list(
    dft = data.frame(suid = c(2, 3), count = c(50, 50)),
    ctus = "a1_b2",
    ctus_orig = c(polyA = "a1", polyB = "b2")
  )
)

tabulate_acrgeo(
  x = xn,
  varname = "var",
  is_var_categorical = FALSE,
  ctus = ctus,
  fun = fun_acrossgeo_numeric
)

tabulate_acrgeo(
  x = xc1,
  varname = "var",
  is_var_categorical = TRUE,
  ctus = ctus,
  fun = fun_acrossgeo_categorical
)

tabulate_acrgeo(
  x = xc1,
  varname = "var",
  is_var_categorical = TRUE,
  ctus = ctus,
  fun = fun_acrossgeo_categorical,
  levels = 0:5
)

tabulate_acrgeo(
  x = xc2,
  varname = "var",
  is_var_categorical = TRUE,
  ctus = ctus,
  fun = fun_acrossgeo_categorical,
  levels = letters[1:5]
)

}
