% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions__GeoTIFFs.R
\name{write_esri_vat}
\alias{write_esri_vat}
\title{Create ESRI value attribute table \code{"VAT"} as \code{".vat.dbf"} file}
\usage{
write_esri_vat(x, filename, rename_cns = NULL)
}
\description{
A \code{"VAT"} contains fields \code{"VALUE"} and \code{"COUNT"};
all field names are restricted to 11 characters and cannot contain periods.
See \link[foreign:write.dbf]{foreign::write.dbf}
}
